requirements.txt
python-dotenv
langchain-groq
langgraph
pandas
numpy
mcp
streamlit
folium
streamlit-folium
langchain-core
joblib
pydantic
shap
requests
scikit-learn
